+++ include/inviwo/core/properties/transferfunctionproperty.h
@@ -43,17 +43,17 @@ namespace inviwo {
 
 class IVW_CORE_API TransferFunctionPropertyObserver : public Observer {
 public:
-    virtual void onMaskChange(const vec2& mask);
-    virtual void onZoomHChange(const vec2& zoomH);
-    virtual void onZoomVChange(const vec2& zoomV);
+    virtual void onMaskChange(const dvec2& mask);
+    virtual void onZoomHChange(const dvec2& zoomH);
+    virtual void onZoomVChange(const dvec2& zoomV);
     virtual void onHistogramModeChange(HistogramMode mode);
 };
 class IVW_CORE_API TransferFunctionPropertyObservable
     : public Observable<TransferFunctionPropertyObserver> {
 protected:
-    virtual void notifyMaskChange(const vec2& mask);
-    virtual void notifyZoomHChange(const vec2& zoomH);
-    virtual void notifyZoomVChange(const vec2& zoomV);
+    virtual void notifyMaskChange(const dvec2& mask);
+    virtual void notifyZoomHChange(const dvec2& zoomH);
+    virtual void notifyZoomVChange(const dvec2& zoomV);
     virtual void notifyHistogramModeChange(HistogramMode mode);
 };
 
@@ -71,8 +71,8 @@ public:
 
     TransferFunctionProperty(
         const std::string& identifier, const std::string& displayName,
-        const TransferFunction& value = TransferFunction({{0.0f, vec4(0.0f, 0.0f, 0.0f, 0.0f)},
-                                                          {1.0f, vec4(1.0f, 1.0f, 1.0f, 1.0f)}}),
+        const TransferFunction& value = TransferFunction({{0.0, vec4(0.0f, 0.0f, 0.0f, 0.0f)},
+                                                          {1.0, vec4(1.0f, 1.0f, 1.0f, 1.0f)}}),
         VolumeInport* volumeInport = nullptr,
         InvalidationLevel invalidationLevel = InvalidationLevel::InvalidOutput,
         PropertySemantics semantics = PropertySemantics::Default);
@@ -87,14 +87,14 @@ public:
     virtual TransferFunctionProperty* clone() const override;
     virtual ~TransferFunctionProperty();
 
-    const vec2 getMask() const;
-    void setMask(float maskMin, float maskMax);
+    const dvec2 getMask() const;
+    void setMask(double maskMin, double maskMax);
 
-    const vec2& getZoomH() const;
-    void setZoomH(float zoomHMin, float zoomHMax);
+    const dvec2& getZoomH() const;
+    void setZoomH(double zoomHMin, double zoomHMax);
 
-    const vec2& getZoomV() const;
-    void setZoomV(float zoomVMin, float zoomVMax);
+    const dvec2& getZoomV() const;
+    void setZoomV(double zoomVMin, double zoomVMax);
 
     void setHistogramMode(HistogramMode type);
     HistogramMode getHistogramMode();
@@ -114,8 +114,8 @@ public:
     virtual void onTFPrimitiveChanged(const TFPrimitive* p) override;
 
 private:
-    ValueWrapper<vec2> zoomH_;
-    ValueWrapper<vec2> zoomV_;
+    ValueWrapper<dvec2> zoomH_;
+    ValueWrapper<dvec2> zoomV_;
     ValueWrapper<HistogramMode> histogramMode_;
 
     VolumeInport* volumeInport_;
